backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# WebSocket support
websockets==12.0

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Google Generative AI SDK (Gemini)
google-generativeai==0.3.2

# Image processing
Pillow==10.2.0

# Async support
aiohttp==3.9.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
